refactor: enum 타입 constants 폴더의 enums 폴더로 분류 (#1)
