+++ user/src/main/java/com/sangyunpark/user/constant/enums/UserType.java
@@ -1,0 +1,5 @@
+package com.sangyunpark.user.constant.enums;
+
+public enum UserType {
+    NORMAL, ADMIN
+}
