--- user/src/main/java/com/sangyunpark/user/domain/vo/UserType.java
@@ -1,5 +1,0 @@
-package com.sangyunpark.user.domain.vo;
-
-public enum UserType {
-    NORMAL, ADMIN
-}
