+++ user/src/main/java/com/sangyunpark/user/presentation/UserController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import static com.sangyunpark.user.constant.ResponseMessages.*;
+import static com.sangyunpark.user.constant.message.ResponseMessages.*;
 
 @RestController
 @RequestMapping("/api/v1/users")
