+++ user/src/test/java/com/sangyunpark/user/application/UserServiceTest.java
@@ -4,9 +4,9 @@ import com.sangyunpark.user.application.mapper.UserMapper;
 import com.sangyunpark.user.domain.dto.request.UserAddressRequestDto;
 import com.sangyunpark.user.domain.dto.request.UserSignupRequestDto;
 import com.sangyunpark.user.domain.entity.User;
-import com.sangyunpark.user.domain.vo.RegisterType;
-import com.sangyunpark.user.domain.vo.UserStatus;
-import com.sangyunpark.user.domain.vo.UserType;
+import com.sangyunpark.user.constant.enums.RegisterType;
+import com.sangyunpark.user.constant.enums.UserStatus;
+import com.sangyunpark.user.constant.enums.UserType;
 import com.sangyunpark.user.exception.UserDuplicateException;
 import com.sangyunpark.user.infrastructure.repository.UserJpaRepository;
 import org.junit.jupiter.api.BeforeEach;
@@ -15,7 +15,7 @@ import org.junit.jupiter.api.Test;
 
 import java.util.Optional;
 
-import static com.sangyunpark.user.constant.ExceptionMessages.EXCEPTION_USER_DUPLICATE;
+import static com.sangyunpark.user.constant.message.ExceptionMessages.EXCEPTION_USER_DUPLICATE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.BDDMockito.given;
