+++ user/src/test/java/com/sangyunpark/user/integration/UserIntegrationTest.java
@@ -2,9 +2,9 @@ package com.sangyunpark.user.integration;
 
 import com.sangyunpark.user.domain.entity.User;
 import com.sangyunpark.user.domain.entity.UserAddress;
-import com.sangyunpark.user.domain.vo.RegisterType;
-import com.sangyunpark.user.domain.vo.UserStatus;
-import com.sangyunpark.user.domain.vo.UserType;
+import com.sangyunpark.user.constant.enums.RegisterType;
+import com.sangyunpark.user.constant.enums.UserStatus;
+import com.sangyunpark.user.constant.enums.UserType;
 import com.sangyunpark.user.infrastructure.repository.UserJpaRepository;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.DisplayName;
@@ -15,8 +15,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
-import static com.sangyunpark.user.constant.ExceptionMessages.EXCEPTION_USER_DUPLICATE;
-import static com.sangyunpark.user.constant.ResponseMessages.*;
+import static com.sangyunpark.user.constant.message.ExceptionMessages.EXCEPTION_USER_DUPLICATE;
+import static com.sangyunpark.user.constant.message.ResponseMessages.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
